Rich text to plaintext descriptions, other ui changes
